docs(sidebarcontent): created a reusable sidebar component
